% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guitar_chords.R
\name{chord_for_frets}
\alias{chord_for_frets}
\title{Chord for Frets}
\usage{
chord_for_frets(frets, tuning = c("e", "a", "d", "g", "b", "e"))
}
\arguments{
\item{frets}{A 6 element vector representing fret positions.}

\item{tuning}{A 6 element vector representing notes of open strings.}
}
\value{
Returns a 6 element vector of note names.
}
\description{
Return the notes for a given set of frets for a specified tuning.
Preserves NAs indicating muted strings.
}
\examples{
chord_for_frets(c(NA, 3, 2, 0, 1, 0))

}
\seealso{
\code{\link{notes_for_frets}}
}

