% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guitar_chords.R
\name{narrow_chord}
\alias{narrow_chord}
\title{Narrow (Filter) Chord Selection}
\usage{
narrow_chord(chords, value, name)
}
\arguments{
\item{chords}{A data frame of chords.}

\item{value}{The value to be filtered.}

\item{name}{The name to be filtered.}
}
\value{
Returns a data frame of chords.
}
\description{
This function filters rows in the chord data frame
by criteria passed in and accounts for nulls.
The nifty bit is - given a full set of chords, you can get chords that
use a given fret or frets.
}

