% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guitar_chords.R
\name{select_chords}
\alias{select_chords}
\title{Select chords by criteria}
\usage{
select_chords(note_name = NULL, chord_name = NULL, string_1_fret = NULL,
  string_2_fret = NULL, string_3_fret = NULL, string_4_fret = NULL,
  string_5_fret = NULL, string_6_fret = NULL)
}
\arguments{
\item{note_name}{Root of the chord.}

\item{chord_name}{Name of the chord.}

\item{string_1_fret}{Fret location on the 1st string.}

\item{string_2_fret}{Fret location on the 2nd string.}

\item{string_3_fret}{Fret location on the 3rd string.}

\item{string_4_fret}{Fret location on the 4th string.}

\item{string_5_fret}{Fret location on the 5th string.}

\item{string_6_fret}{Fret location on the 6th string.}
}
\value{
Returns a data frame of chords.
}
\description{
Select chords by criteria
}
\examples{
select_chords(chord_name = 'Major', string_5_fret = 3)
select_chords(note='d')
}

